# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "https://localhost:8443",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/users/{userId}/tweets": {
      "get": {
        "tags": [
          "timeline-controller"
        ],
        "operationId": "userTweets",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTweetView"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tweets/{tweetId}/replies": {
      "get": {
        "tags": [
          "timeline-controller"
        ],
        "operationId": "replies",
        "parameters": [
          {
            "name": "tweetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTweetView"
                }
              }
            }
          }
        }
      }
    },
    "/v1/timeline": {
      "get": {
        "tags": [
          "timeline-controller"
        ],
        "operationId": "homeTimeline",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTweetView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageTweetView": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TweetView"
            }
          },
          "nextCursor": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TweetView": {
        "type": "object",
        "properties": {
          "tweetId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "inReplyToTweetId": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "repostCount": {
            "type": "integer",
            "format": "int64"
          },
          "replyCount": {
            "type": "integer",
            "format": "int64"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "repostedByMe": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}